fix: 배경 컴포넌트 img -> div 변경
